This should fix #3909
We created empty branches when we attempted to split a QB node
with ways in it and none of the ways could fit into a child.
The children were allocated before we knew that we had something
to put in them, and we ended up with an empty array of children.
That triggered this check.

So, defer child array allocation until we really need it.



git-svn-id: https://josm.openstreetmap.de/svn/trunk@2442 0c6e7542-c601-0410-84e7-c038aed88b3b