fix #7671 - Show last N used tags in "Add key/value" dialog for selecting with a single click
LRU tags are shown in reverse order (most recent first).

This can be customized by setting the property `properties.recently-added-tags` to an integer between 1 and 9.
This can be disabled by setting the same property to a number lesser or equal to 0.

Keyboard shortcuts Ctrl+1 to Ctrl+9 are available in the context of the "Add property" dialog only.

git-svn-id: https://josm.openstreetmap.de/svn/trunk@5383 0c6e7542-c601-0410-84e7-c038aed88b3b